feat(monorepos): Add BUILDKITE_ANALYTICS_LOCATION_PREFIX var
This adds the ability to set a `BUILDKITE_ANALYTICS_LOCATION_PREFIX`
environment variable that will prefix all test paths with the given path. This
is especially useful when using the test collector in a monorepo setting,
which has multiple packages being tested in tandem as separate CI builds.

This moves the `new CI` call up as a class property `this._testEnv`, which
is used in both `onTestResult` and `uploadTestResults`.