Attempt to log network calls seperated by file
Test files are run in parallel by multiple jest workers, but each test
file will only be run by a single worker, so we can group tests
belonging to a single file in a temporary storage before adding to the
global test results array.

Some data (like the filename of the test file), is only available in the
after test file hook, so we need to add this back into our temporary
results.

We can also assign each test a tracer using the hooks by storing the
tracer in the test configuration which persists between jest hooks.

This solution actually works when jest is set to `--runInBand` but it
does not work when jest is using multiple workers, I suspect this is due
to the context of the network setup, when jest is running in band the
context is the same as the test execution, but when there are workers,
it's run in a separate process, so we can't hook into network calls
unless we can figure out a way to run the network setup in the same
context as the workers.

The `setupFilesAfterEnv` script IS run in the same context as the
workers which is great, but I don't know of a way of sending information
from the test hooks to the reporter hooks.