Enable JUnit output from Jest
I'd like to have a demo suite of an open source codebase we can use
while building public suites for TA. Using one of our collectors as an
example seemed like the easiest place to start, but I expect it will
change in future.

In theory it might be possible to use a recursive solution and use the
Javascript Collector for the Javascript Collector (using this repo
against itself); however, that seemed unnecessarily complicated and
risky. I felt it was easier to use a standard JUnit formatter. The
results will need to be uploaded using the collector plugin.